services/ai-orchestrator/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.34
langgraph==0.0.40
openai==1.14.3
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.6.4
redis==5.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
alembic==1.13.1
googlemaps==4.10.0
sentry-sdk==1.44.0
celery==5.3.6
tenacity==8.2.3
beautifulsoup4==4.12.3
aiohttp==3.9.3
